embed privacy policy in app
- create privacy policy activity
- refactor map to web view activity to share common behavior
- open Privacy Policy from the app menu
- contributes to https://github.com/niccokunzmann/mundraub-android/issues/205
