+++ app/src/main/java/eu/quelltext/mundraub/activities/MundraubBaseActivity.java
@@ -127,6 +127,9 @@ public class MundraubBaseActivity extends AppCompatActivity implements Logger.Lo
             case R.id.item_rules:
                 menuOpenCommunityCodex();
                 return true;
+            case R.id.item_privacy:
+                menuOpenPrivacyPolicy();
+                return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
@@ -149,6 +152,11 @@ public class MundraubBaseActivity extends AppCompatActivity implements Logger.Lo
 
     protected void menuOpenMap() {
         Intent intent = new Intent(this, ShowPlantsActivity.class);
+        this.startActivity(intent);
+    }
+
+    protected void menuOpenPrivacyPolicy() {
+        Intent intent = new Intent(this, PrivacyPolicyActivity.class);
         this.startActivity(intent);
     }
